Simplify config file handling
Instead of four arguments: a log, tradlog, config and depot file the trader now
accepts two arguments logfile and data_dir. The data_dir is supposed
to contain a depot file and a config, where the trader reads from. A
tradlog file will be created if it does not exist.

The trader will log to the logfile, since these file tend to be large,
they can be located elsewhere.